Support image trait when boot from volume
When you boot from volume, Nova returns nothing for the image id,
and so we can't test for the trait against that.

Nova behind the scenes looks at the boot volume and the trait is
inhertied on it from the original image.

This commit adds support for attempting to look up the trait from
any attached volumes if an image id wasn't provided.

Change-Id: Ib608852fa71105aa9b197b3250d8527e0d56ddb0