Don't use stale vault client
Turns out that the k8s token we use for Vault expires after 1 hour.

Due to caching, we weren't creating a new vault client, and so
eventually the token would expire, but the original vault client
would continue to use the expired token.

This commit just instantiates a new client each time the secrets
are requested (outside of our 60s cache)

Change-Id: I1cb39c8c4fe61a6550c523e14e0570c3f95faa36